+++ spec/chewy/index/adapter/active_record_spec.rb
@@ -1,5 +1,11 @@
 require 'spec_helper'
 
+RawCity = Struct.new(:id) do
+  def rating
+    id * 10
+  end
+end
+
 describe Chewy::Index::Adapter::ActiveRecord, :active_record do
   before do
     stub_model(:city)
@@ -569,6 +575,26 @@ describe Chewy::Index::Adapter::ActiveRecord, :active_record do
             city_ids, _index: 'users', scope: City.where(country_id: 1), users: {scope: -> { where(country_id: 0) }}
           )
         ).to eq(cities.first(2) + [nil])
+      end
+    end
+
+    context 'with raw_import option' do
+      subject { described_class.new(City) }
+
+      let!(:cities) { Array.new(3) { |i| City.create!(rating: i / 2) } }
+      let(:city_ids) { cities.map(&:id) }
+
+      let(:raw_import) { ->(hash) { RawCity.new(hash['id']) } }
+      it 'uses the custom loader' do
+        raw_cities = subject.load(city_ids, _index: 'cities', raw_import: raw_import).map do |c|
+          {id: c.id, rating: c.rating}
+        end
+
+        expect(raw_cities).to eq([
+          {id: 1, rating: 10},
+          {id: 2, rating: 20},
+          {id: 3, rating: 30}
+        ])
       end
     end
   end